+++ src/Sulu/Bundle/RouteBundle/Exception/MissingClassMappingConfigurationException.php
@@ -16,8 +16,50 @@ namespace Sulu\Bundle\RouteBundle\Exception;
  */
 class MissingClassMappingConfigurationException extends \Exception
 {
-    public function __construct($className)
+    /**
+     * @var string
+     */
+    private $className;
+
+    /**
+     * @var string[]
+     */
+    private $available;
+
+    /**
+     * @param string $className
+     * @param string[] $available
+     */
+    public function __construct($className, array $available)
     {
-        parent::__construct(sprintf('Missing class mapping configuration for "%s"', $className));
+        parent::__construct(
+            sprintf(
+                'Missing class mapping configuration for "%s". Available classes: ["%s"]',
+                $className,
+                implode('", "', $available)
+            )
+        );
+        $this->className = $className;
+        $this->available = $available;
+    }
+
+    /**
+     * Returns requested class-name.
+     *
+     * @return string
+     */
+    public function getClassName()
+    {
+        return $this->className;
+    }
+
+    /**
+     * Returns available configurations.
+     *
+     * @return string[]
+     */
+    public function getAvailable()
+    {
+        return $this->available;
     }
 }